requirements.txt
fastapi
uvicorn
langgraph
pytrends
google-api-python-client
python-dotenv
google-genai
